+++ addons/crates_c/CfgVehicles.hpp
@@ -1,0 +1,26 @@
+class CfgVehicles {
+	class 16aa_crate_empty;
+	class 16aa_crate_empty_artillery;
+	class 16aa_crate_empty_csw;
+	class 16aa_crate_empty_main;
+	class 16aa_crate_empty_medical;
+	class 16aa_crate_empty_launchers;
+	class 16aa_crate_empty_platoon;
+	class 16aa_crate_empty_wooden_large;
+	class 16aa_crate_empty_wooden_small;
+	class Land_Cargo20_military_green_F;
+	class Cargo_base_F;
+	#include "Crates\Crate_Ammo.hpp"
+	#include "Crates\Crate_CSW.hpp"
+	#include "Crates\Crate_Equipment.hpp"
+	#include "Crates\Crate_Flight.hpp"
+	#include "Crates\Crate_FSG.hpp"
+	#include "Crates\Crate_Main.hpp"
+	#include "Crates\Crate_Medical.hpp"
+	#include "Crates\Crate_Ordnance.hpp"
+	#include "Crates\Crate_Platoon.hpp"
+	#include "Crates\Crate_Radio.hpp"
+	#include "Crates\Crate_RHA.hpp"
+	#include "Crates\Crate_Section.hpp"
+	#include "Crates\Crate_Training.hpp"
+};
