+++ addons/crates_c/Crates/Crate_Equipment.hpp
@@ -1,17 +1,4 @@
-class CfgPatches {
-	class 16AA_crate_equipment {
-		requiredaddons[] = {"16aa_crates_a3"};
-		requiredversion = 0.1;
-		units[] = {"16AA_crate_equipment"};
-		weapons[] = {};
-		magazines[] = {};
-	};
-};
-
-class CfgVehicles {
-	class 16aa_crate_empty;	// External class reference
-
-	class 16AA_crate_equipment : 16aa_crate_empty {
+	class 16AA_crate_equipment: 16aa_crate_empty {
 		author = "Adams & Beazley";
 		displayName = "Equipment";
 		vehicleClass = "16aa_Crates";
@@ -31,4 +18,3 @@ class CfgVehicles {
         	MACRO_ADDBACKPACK(16aa_B_Bergen_MTP_Rifleman_L_A,20);
         };
     };
-};
