+++ addons/crates_c/Crates/Crate_FSG.hpp
@@ -1,17 +1,4 @@
-class CfgPatches {
-	class 16AA_crate_fsg {
-		requiredaddons[] = {"16aa_crates_a3"};
-		requiredversion = 0.1;
-		units[] = {"16AA_crate_fsg"};
-		weapons[] = {};
-		magazines[] = {};
-	};
-};
-
-class CfgVehicles {
-	class 16aa_crate_empty_main;	// External class reference
-
-	class 16AA_crate_fsg : 16aa_crate_empty_main {
+	class 16AA_crate_fsg: 16aa_crate_empty_main {
 		author = "Adams & Beazley";
 		displayName = "FSG";
 		vehicleClass = "16aa_Crates";
@@ -54,4 +41,3 @@ class CfgVehicles {
         	MACRO_ADDBACKPACK(16aa_B_Bergen_MTP_Radio_L_B,2);
         };
 	};
-};
