+++ addons/crates_c/config.cpp
@@ -10,17 +10,129 @@ class CfgPatches {
         authorUrl = "";
         VERSION_CONFIG;
     };
+    class 16AA_crate_ammo {
+    	units[] = {
+			"16AA_container_ammo",
+			"16AA_container_ammo_artillery",
+			"16AA_crate_csw_ammo",
+			"16AA_crate_artillery_105",
+			"16AA_crate_artillery_81",
+			"16AA_crate_ammo_556_stanag",
+			"16AA_crate_ammo_556_link",
+			"16AA_crate_ammo_762_stanag",
+			"16AA_crate_ammo_762_link",
+			"16AA_crate_ammo_12ga",
+			"16AA_crate_ammo_grenades_smoke",
+			"16AA_crate_ammo_grenades_frag",
+			"16AA_crate_ammo_grenades_flashbang",
+			"16AA_crate_ammo_grenades_chemlight",
+			"16AA_crate_ammo_grenades_40mm",
+			"16AA_crate_ammo_explosives",
+			"16AA_crate_ammo_launcher",
+			"16AA_crate_ammo_misc"
+		};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_csw {
+		units[] = {"16AA_crate_csw"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_equipment {
+		units[] = {"16AA_crate_equipment"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_flight {
+		units[] = {"16AA_crate_flight"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_fsg {
+		units[] = {"16AA_crate_fsg"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_main {
+		units[] = {"16AA_crate_main"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_medical {
+		units[] = {"16AA_crate_medical"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16aa_crate_ordnance {
+		units[] = {"16AA_crate_ordnance"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_platoon {
+		units[] = {"16AA_crate_platoon"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_radio {
+		units[] = {"16AA_crate_radio"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_rha {
+		units[] = {"16AA_crate_rha"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crate_section {
+		units[] = {"16AA_crate_section"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
+	class 16AA_crates_training
+	{
+		units[] = {"16AA_crate_training_main","16AA_crate_training_section","16AA_crate_csw_training"};
+		requiredaddons[] = {"LSR_common","16aa_crates_a3"};
+		requiredversion = REQUIRED_VERSION;
+		author[] = {"Grey"};
+        authorUrl = "";
+        VERSION_CONFIG;
+	};
 };
-#include "Crates\Crate_Ammo.hpp"
-#include "Crates\Crate_CSW.hpp"
-#include "Crates\Crate_Equipment.hpp"
-#include "Crates\Crate_Flight.hpp"
-#include "Crates\Crate_FSG.hpp"
-#include "Crates\Crate_Main.hpp"
-#include "Crates\Crate_Medical.hpp"
-#include "Crates\Crate_Ordnance.hpp"
-#include "Crates\Crate_Platoon.hpp"
-#include "Crates\Crate_Radio.hpp"
-#include "Crates\Crate_RHA.hpp"
-#include "Crates\Crate_Section.hpp"
-#include "Crates\Crate_Training.hpp"
+#include "CfgVehicles.hpp"
