+++ addons/main/script_macros.hpp
@@ -67,7 +67,10 @@
   magazine = #MAGAZINE; \
   count = COUNT; \
 }
-
+#define MACRO_ADDBACKPACK(BACKPACK,COUNT) class _xx_##BACKPACK { \
+  backpack = #BACKPACK; \
+  count = COUNT; \
+}
 
 #ifdef DISABLE_COMPILE_CACHE
     #define PREP(fncName) DFUNC(fncName) = compile preprocessFileLineNumbers QUOTE(PATHTOF(functions\DOUBLES(fnc,fncName).sqf))
